web/requirements.txt
passlib
PyJWT
flask
requests
python-dotenv
flask_wtf
email_validator
argon2-cffi==16.3.0